backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Validation & Serialization
pydantic==2.5.0
email-validator==2.1.0

# HTTP Client & External APIs
httpx==0.25.2
requests==2.31.0

# Background Jobs & Caching
redis==5.0.1
celery==5.3.4

# File Storage & Image Processing
boto3==1.34.0
Pillow==10.1.0
python-magic==0.4.27

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Utilities
python-slugify==8.0.1
jinja2==3.1.2
python-dateutil==2.8.2
pytz==2023.3

# Optional: For better performance
orjson==3.9.10

# Optional: For development
watchfiles==0.21.0

# Third-party API Integrations
urllib3==2.1.0
cryptography==41.0.7

# OAuth 2.0 and Authentication for APIs
authlib==1.2.1
itsdangerous==2.1.2

# Two-Factor Authentication
pyotp==2.9.0
qrcode[pil]==7.4.2

# Additional HTTP and API utilities
aiohttp==3.9.1
backoff==2.2.1

# Rate limiting and retry logic
tenacity==8.2.3

# JSON Web Tokens for Shopify
PyJWT==2.8.0